More renamining:
long_blah -> z_blah

long_blah now takes up to 53 bit operands on 64 it machine or 31 bits on a 32 bit machine

long_blah_64 provides 63 bits on a 64 bit machine and 31 bits on a 32 bit machine

long_blah2 chooses between long_blah and long_blah_64 and picks whichever will give the maximum number of bits available on the given architecture (obviously picking the faster of the two on the 32 bit architecture)

It is now only necessary to use long_blah and long_blah2 functions. The first will be faster on a 64 bit machine but only give 53 bits. The two functions will be identical on a 32 bit machine.


git-svn-id: https://flint.svn.sourceforge.net/svnroot/flint/trunk@983 2531bd8e-7522-0410-9704-d59f193d18de