Added modular gcd function to fmpz_poly, plus test code.
Fixed some bugs in the classical division algorithms in zmod_poly.c

Did some work on the makefile to make things compile.

Fixed the test code for the subresultant gcd function.

Made a few minor modifications to the memory-manager to the debugging realloc function.

Updated todo.txt.

Added some functions to fmpz including a CRT function and a cmpabs function.

git-svn-id: https://flint.svn.sourceforge.net/svnroot/flint/trunk@1169 2531bd8e-7522-0410-9704-d59f193d18de